encryption_tool/requirements.txt
pycryptodome
matplotlib
numpy
pillow
